Introduce RabbitMQ
Classifying Tweets with our classifier is not fast enough to do in real
time and keep up with Twitter's stream, so put the Tweets in a queue to
be classified by a separate service.

This classification happens at the same time as streaming, so some
Tweets in the database will have `null` set as their `sentiment` until
the classifier has caught up with the Twitter queue.