Use Ajax to load new Tweets in the result view
Use Ajax to send a request to the controller, which renders any Tweets
that are in the database but not yet displayed on the view, and returns
the rendered HTML.

This HTML is then appended to the DOM element containing the list of
Tweets.

Currently sends a new Ajax request every 5000ms but this might need to
be tweaked. Each request seems to take around 1000ms to complete so the
interval between new requests could be reduced.